refactor: split status in multiple files
